Extract code to smaller functions. Rename module.
