+++ tests/test_emissions.py
@@ -1,7 +1,7 @@
 import unittest
 
-from codecarbon.utils.emissions import Emissions
-from codecarbon.units import Energy
+from codecarbon.core.emissions import Emissions
+from codecarbon.core.units import Energy
 from codecarbon.external.geography import CloudMetadata, GeoMetadata
 
 from tests.testutils import get_test_data_source
